requirements.txt
Flask~=2.0.2
requests~=2.26.0
urllib3~=1.26.7
dnspython~=2.1.0
setuptools~=57.0.0
flask-eureka-client~=1.0.1